% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/palingdensiteit.R
\name{calcMultimarkPopSizeIterative}
\alias{calcMultimarkPopSizeIterative}
\title{Schatting populatie N voor multiple capture mark recapture}
\usage{
calcMultimarkPopSizeIterative(data, colnames = c("vangst", "hervangst"),
  max_iter = 100, multiplier = 2, plot = TRUE, return = c("N",
  "seN", "delta_N", "lcl", "ucl"))
}
\arguments{
\item{data}{dataset die de kolommen bevat die onder 'colnames' staan}

\item{colnames}{de kolomnamen die overeenkomen met vangst en hervangst respectievelijk}

\item{max_iter}{maximum aantal iteraties dat de procedure uitvoert vooraleer op te geven}

\item{multiplier}{vermeningvuldigingsfactor voor de startschatting die standaard op het aantal markeringen ingesteld staat}

\item{plot}{toon een plot die de evolutie van de schatting N en de delta_N toont, bij convergentie zal dat met een vlakkel lijn eindigen}

\item{return}{Welke velden moeten teruggegeven worden. Je kan 1 of meerdere uit location, year, N, seN en delta_N kiezen}
}
\value{
data.frame met 1 rij met de kolommen gespecifieerd in de parameter 'return'
}
\description{
Iteratieve methode om populatie N in te schatten voor meerdere vangstoccasies waar iedere vis die niet nog niet gemerkt was van ee markering voorzien wordt
}
\examples{
{
df <- data.frame(vangst = c(36, 31, 25, 34), hervangst = c( 0,  4,  7, 15))
PopEstim <-
calcMultimarkPopSizeIterative(data = df, multiplier = 1, plot = TRUE)
PopEstim
}
}
