% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/palingdensiteit.R
\name{calcN}
\alias{calcN}
\title{Berekening van populatiegrootte (N) door iteratieve schatting voor multiple mark-recapture}
\usage{
calcN(N, catch, recapture)
}
\arguments{
\item{N}{Vorige populatieschatting}

\item{catch}{de totale vangsten voor iedere meetoccasie}

\item{recapture}{de hoeveelheid hervangsten voor iedere meetoccasie (de eerste is standaard 0)}
}
\value{
vector die de nieuwe populatieschatting bevat alsook de berekende delta
}
\description{
Bereken een nieuwe N en de delta_n op basis van de formules zoals beschreven in het handboek Ricker (1968). Dit is meer specifiek enkel geldig voor de schatting van populatiegrootte waarbij op verschillende momenten kort na elkaar de dieren gevangen worden, en alle nog-niet-gemerkte dieren gemerkt worden zodat ze herkenbaar zijn bij de volgende vangst. De basis voor deze berekening is dat het totaal niet-gemerkte dieren in de populatie gelijk gesteld wordt aan de kans dat een dier tijdens geen enkele vangst gemerkt wordt. Zowel N als delta_n zullen convergeren naar een vaste waarde en dan stopt het iteratieve proces. Deze N wordt de populatieschatting en delta_n wordt gebruikt om de onzekerheid in de schatting te bepalen, namelijk SE(N) = sqrt((N-U)/delta_n). De kans dat een dier in geen enkele vangst gevangen wordt is het product van de kansen dat ze tijdens een bepaalde vangstoccasie gevangen worden. PN(1,2,...,X) = PN(1)PN(2)...PN(X) met PN(i) = (N - catch_i)/N. Voor de berekening van delta_n wordt de geschatte populatie vergeleken met de waarde van de formule waarbij we N-1 invullen waar N voorkomt. Indien de schatting voor N-1 nagenoeg 1 minder is dan de schatting voor N, dan wil dit zeggen dat de formule even goed geldt voor een andere N waarde dan de geschatte waarde, en is er dus een grotere onzekerheid op de schatting, aangezien andere N waarden even goed werken in de formule. De nieuwe inschatting voor N wordt bekomen door een soort interpolatie. Er wordt gekeken hoeveel N * PN afwijkt van N - totalmarks en de nieuwe inschatting voor N is dan N + dit verschil (wat ook negatief kan zijn) gedeeld door de delta_n. De deling door delta_n zorgt dat er veel sneller convergentie komt, anders wordt dat een zeer traag assymptotisch proces waarvoor heel veel iteraties nodig zijn
}
