% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/mark_lifehistory_barker.R
\name{mark_lifehistory_barker}
\alias{mark_lifehistory_barker}
\title{Omzetten naar Barker LD (Chapter 9.5 in mark handleiding)}
\usage{
mark_lifehistory_barker(data, colnames = c("id", "date", "event",
  "group"), names_event = c("ringing", "liverecovery", "sighting",
  "deadrecovery"), dates_event)
}
\arguments{
\item{data}{dataset met de nodige gegevens, met tenminste een identificatiekolom, een datumkolom (numeriek), een kolom met de state (bijvoorbeeld leeftijd bij ringen) en een kolom die het event beschrijft (ring aanbrengen, levend gezien, dood gevonden)}

\item{colnames}{om de kolommen te herkenen in de functie moet je deze in volgorde identificeren, de eerste kolomnaam moet de id bevatten, de tweede een numeriek tijdstip, de derde een event, en de volgende kolommen kunnen gebruikt worden voor groeperingsvariabelen}

\item{names_event}{hoe de events in de data genoemd zijn, de eerste komt overeen met een ringevent, de tweede met een levende recovery tijdens een vangstevent, de derde zijn de levend gevangen dieren tussen events door, de vierde zijn de doodgevonden dieren}

\item{dates_event}{Voor het Barker model wordt onderscheid gemaakt tussen de effectieve vangst- en ringevents, en de levend- en doodgevonden dieren tijdens het interval tussen 2 vangstevents.}
}
\value{
data.frame with life history, containing columns Ch, freq and group
}
\description{
Omzetten naar Barker LD (Chapter 9.5 in mark handleiding)
}
\examples{
{
ringing <- data.frame(id = 1:20, date = c(rep(1,10), rep(2,10)), 
                      group = rep(rep(0:1, 5), 2), event = "ringing")
period2 <- data.frame(id = c(1,2,3), date = 2, group = c(1,1,1), event = "liverecovery")
period2a <- data.frame(id = c(4,5), date = 2, group = c(1,1), event = "sighting")
period2b <- data.frame(id = c(4, 20), date = 2.5, group = c(2,1) ,event = "deadrecovery")
period3 <- data.frame(id = c(1,3,5,12,15), date = 3, group = c(2,2,2,1,1), event = "liverecovery")
testdata <- rbind(ringing, period2, period2a, period2b, period3)
lddata <- mark_lifehistory_barker(testdata, dates_event = 1:3) 
print(lddata)
}
}
