% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/sim_multimark_density.R
\name{sim_multimark_density}
\alias{sim_multimark_density}
\title{Simuleer een dataset met vangstgegevens om de palingdensiteitsberekening uit te voeren}
\usage{
sim_multimark_density(Npop = 200, S = 30, Ndates = 5, Nsim = 200,
  verbose = TRUE)
}
\arguments{
\item{Npop}{De echte populatiegrootte}

\item{S}{Hoeveel vissen worden er gemiddeld per vangst gevangen}

\item{Ndates}{Hoeveel dagen naëen vis je om de populatieschatting te verbeteren}

\item{Nsim}{Hoeveel keer moet de routine opnieuw lopen om een zicht te krijgen op de variabiliteit van de schattingen}

\item{verbose}{indien TRUE toon op het scherm extra informatie over het verloop van het proces}
}
\value{
data.frame met per Nsim, per Ndates de vangstgrootte, de recaptures, en de geschatte populatie volgens de methode op basis van de informatie tot de respectievelijke vangstdag
}
\description{
Simuleer een dataset met vangstgegevens om de palingdensiteitsberekening uit te voeren
}
\examples{
schattingen <- sim_multimark_density(Npop = 200, S = 30, Nsim = 100)
head(schattingen)
hist(schattingen[schattingen$day == 5, "N_estim"], main = "", xlab = "estimation of N")
abline(v = 200, col = "red", lwd = 3)

}
